Update experimental settings behavior
Previously, experimental settings were expected to reset to their default values when experimental features were disabled, but this behavior was not correctly implemented.
Now, the values of experimental settings will be preserved but will not be applied if experimental features are disabled.
Also, previous statements describing the now-outdated behavior have been removed from the attachments.