Introduce system allowing to add L&F via plugins
Users can now add third-party Java Swing Look and Feels (L&Fs) to the "JSON Editor" via plguins. (Currently, this only works for directory plugins.)
To do this, users should create a folder named `look_and_feel` in the root directory of the plugin, and place the JAR file that contains Look and Feel classes within that folder. Additionally, a `look_and_feel.json` file muse be created in the same folder to define which classes are Look and Feel classes and the names of these Look and Feels.
After these L&Fs are loaded by the "JSON Editor", users can select their preferred style in the setting dialog. Note that this feature is experimental features.