Update language file format
There are 2 supported formats:
1. A JSON object containing a `"translations"` key whose value is a JSON object, along with optional metadata.
2. A JSON object representing the value of the `"translations"` key in format 1.

If the `"translations"` key is present and its value is a JSON object, the file is treated as format 1. This format allows specifying alternative languages. When a translation key is missing in the current language, the "JSON Editor" will try to retrieve it from the alternatives before retreving it from the default langauge.